always respect OnlyShowIn and NotShowIn
And also respect case.  The Autostart spec does not mention
case but the Desktop Entry spec says that "Case is significant
everywhere in the file".

The Autostart spec says:

  The OnlyShownIn entry may contain a list of strings identifying the
  desktop environments that MUST autostart this application, all other
  desktop environments MUST NOT autostart this application.

  The NotShownIn entry may contain a list of strings identifying the
  desktop environments that MUST NOT autostart this application, all
  other desktop environments MUST autostart this application.

So we get no say on the matter; if we're not one of the DEs in
`OnlyShowIn' we are not to execute the application.